[TEST] Exposing matlab comparison function
