+++ pyls/tests/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['compare_matlab_result']
+
+from .test_matlab_comp import compare_matlab_result
